MAINT: interpolate: move RGI to a separate file
Also move tests of RGI / interpn to test_rgi.py
