+++ scipy/interpolate/__init__.py
@@ -183,6 +183,8 @@ from ._bsplines import *
 
 from ._pade import *
 
+from ._rgi import *
+
 # Deprecated namespaces, to be removed in v2.0.0
 from . import fitpack, fitpack2, interpolate, ndgriddata, polyint, rbf
