+++ scipy/interpolate/tests/meson.build
@@ -11,7 +11,8 @@ py3.install_sources([
     'test_polyint.py',
     'test_rbf.py',
     'test_rbfinterp.py',
-    'test_regression.py'
+    'test_regression.py',
+    'test_rgi.py'
   ],
   pure: false,
   subdir: 'scipy/interpolate/tests'
